Moving HoodieCatalystExpressionUtils to org.apache.spark.sql
